Refactored the internal mouse-handling code to use floating point coords. This shouldn't affect much user code, but a few methods in MouseInputSource have now changed to use Point<float> rather than Point<int>.

+++ modules/juce_audio_plugin_client/utility/juce_FakeMouseMoveGenerator.h
@@ -51,7 +51,7 @@ public:
                 if (Component* const comp = Desktop::getInstance().findComponentAt (screenPos))
                     if (ComponentPeer* const peer = comp->getPeer())
                         if (! peer->isFocused())
-                            peer->handleMouseEvent (0, peer->globalToLocal (screenPos), mods, Time::currentTimeMillis());
+                            peer->handleMouseEvent (0, peer->globalToLocal (screenPos.toFloat()), mods, Time::currentTimeMillis());
         }
     }
